Highlight areas which are requiring triaging attention.
When triaging bugs, one might be interested into filling the fields
which are not triage properly. This modification will highlight with a
small red dashed border the fields which are inconsistent.

Highlights are added for the priority, severity fields as well as for
the blocks field only if the bug is in one of the JavaScript components,
and if Bug SpiderMonkey is not listed in the transitive list of blocked
bugs.

Additional changes:
 - Improve the README.md
 - Add a flake.nix to redo the zip command.
 - Add example-01.png to be included in the README.
 - Add .gitignore to skip the generated artifacts.