Add searchfox links for gdb stack traces.
When detecting gdb stack traces, one would frequently have to
investigate the functions at the given location. To make this task
simpler, this change mutate the content of code tags which are
containing GDB stack traces, and annotate the paths to be links to
searchfox.

Identifying paths and replacing them with links will help with
adding an iframe to preview the code without leaving bugzilla (#6).